pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rocksnitch"
version = "0.1.0"
description = "Mars Perseverance Navcam rock detection >=10 cm, stereo + monocular fusion"
requires-python = ">=3.9"
authors = [{ name = "oms", email = "om.sanan007@gmail.com" }]
readme = "README.md"
license = { text = "Apache-2.0" }

dependencies = [
    "numpy>=1.26",
    "scipy>=1.11",
    "opencv-python>=4.9",
    "pillow>=10",
    "pyarrow>=15",
    "pydantic>=2.6",
    "typer>=0.12",
    "rich>=13.7",
    "einops>=0.7",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "scikit-image>=0.22",
    "scikit-learn>=1.4",
]

[project.optional-dependencies]
gpu = [
    "torch>=2.6",
    "torchvision>=0.21",
    "transformers>=4.44",
    "accelerate>=0.33",
    "huggingface-hub>=0.24",
    "timm>=1.0.7",
    "tensorboard>=2.16",
]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "pytest-xdist>=3.6",
    "ruff>=0.5",
    "black>=24",
    "mypy>=1.10",
    "types-PyYAML",
    "types-tqdm",
]
all = ["rocksnitch[gpu,dev]"]

[project.scripts]
rock-snitch = "rocksnitch.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
rocksnitch = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "gpu: requires a CUDA GPU (skipped when unavailable)",
    "models: requires downloaded model weights (skipped when absent)",
    "slow: slow tests (>5s)",
    "integration: cross-module integration tests",
]
filterwarnings = [
    "ignore::DeprecationWarning:torch.*",
    "ignore::UserWarning:torchvision.*",
]

[tool.ruff]
line-length = 100
target-version = "py39"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501", "B008", "N803", "N806", "N812", "RUF012"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
ignore_missing_imports = true
